requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
pip==25.1.1
protobuf==5.29.4
setuptools==80.1.0
wheel==0.45.1